Add game test.
