+++ EngineTest/Baseline/StringLexer.txt
@@ -1,2 +1,2 @@
-(1,0): Variable: $varName
+ (1,0): Variable: $varName
 None
