--- EngineTest/TestFiles/ParserTest.txt
@@ -1,42 +1,0 @@
-include "ParserTest-Include.txt";
-
-item frontPorch;
-item frontHall;
-item backHall;
-item laundryRoom;
-item kitchen;
-item diningRoom;
-item livingRoom;
-
-game
-{
-    player.Location = frontHall;
-
-    frontPorch.Description = "You are on a front porch.";
-    frontHall.Description = "You are in a small entryway.";
-    backHall.Description = "You are in a back hallway.";
-    laundryRoom.Description = "You are in a laundry room.";
-    kitchen.Description = "You are in a kitchen.";
-    diningRoom.Description = "You are in a dining room.";
-    livingRoom.Description = "You are in a living room.";
-
-    var $frontDoor = NewDoorItem(
-        frontPorch,
-        frontHall,
-        Direction.North,
-        "red",
-        "door",
-        DoorState.Locked
-        );
-
-    NewOpening(frontHall, backHall, Direction.North);
-    NewClosedDoor(backHall, laundryRoom, Direction.North);
-    NewOpening(backHall, kitchen, Direction.East);
-    NewOpening(kitchen, diningRoom, Direction.East);
-    NewOpening(diningRoom, livingRoom, Direction.South);
-    NewOpening(livingRoom, frontHall, Direction.West);
-
-    $frontDoor.Key = NewKey("brass", "key", laundryRoom);
-
-    Look();
-}
